Add files via upload
This chip is used as the reservoir for gel solution when Photo-ExM gel chemistry is used. Gel solution is pipetted separately into each well of the chip, then the hiExM device is carefully lined up with the chip and lowered onto it such that each post of the device is immersed in a small volume of the gel soluiton.